Codify interaction between _rawdata, _cframe and attributes
Previously interactions between the subject elements were assumed;
create constructs that explicitly layout how these interactions
are supposed to happen.

In particular:

* User only updates attributes using either dicts or direct
  property manipulation. Translation occurs using newly created
  method _set_attr_from_userprms
* Once attrs are set, they are conveyed to the _cframe, using
  method _set_cframe_from_attrs. The _cframe acts as an in-memory
  cache of the OAG's streams. The _cframe may is redundant for
  old-style OAGs, but will come into its own once new-style OAGs
  are implemented.
* Once the OAGs _cframe is created, it can be either persisted to
  the database, or refreshed from the database using
  _refresh_from_cursor. _cframe results can be propagated to OAG
  attributes via a call to _set_attrs_from_cframe.

Keep in mind that all of this works only if the user carefully
keeps track of the internal state of the _cframe. This burden on
the user will be alleviated by new-style OAGs.

Also rename various methods (__exec_query particularly) so that
they can be inherited.