Add forward lookup capability to OAGs
Analyze foreign keys on other OAGs to find references to
current OAG, and create attributes on current OAG to map to
the other OAGs.

Centralize a lot of SQL in SQL property.

Keep track of oagprops programmatically added to an OAG class so that
they can be reverted when other instances of the OAG are
instantiated. Yes, this has implications for OAG parallelism, but
we're talking about python here -- I think this is a reasonable
compromise for now.