More careful handling of OAGs during assignments
Previous iterations of the oagprop decorator were returning new OAGs
instead of references to the OAG actually being assigned to the graph, leading
to unnecessary object creation and making oagcache management impossible at a
graph level.

Fix this unnecessary duplication and add tests to ensure cache coherence and
correctness.

(Further consolidation of this code is likely)

Other changes of note:

* Clean up extkey handling in oagprop
* Hash based logging to clarify rpc nodes
* Remove unused _initdone class property