Modifications to allow rpc to work for in-memory OAGs
Involves relaxing some assumptions around the presence of ids.

This is very, very shaky code. It passes the functional test, but
I am not at all happy about how the code is structured. Longer
term we need to introduce an in-memory mode with the option to
transition to dbpersist mode.