Implement oagprop profiles for OAGs
Problem Statement

Setting oagprops at the class level results in oagprop crosstalk across
objects in the same process. A solution is needed to make sure that an
instance of a class can only access its own oagprops and not clobber
those belonging to other instances.

Solution

Modify the object creation and deletion process to maintain oagprops in a
multitenant structure owned by the class instead of having class instances
do so on an ad-hoc basis.

Context switch functions reset and hydrate oagprops on the class for
access by its instances.

Known Issues:

The multitenant structure is prone to memory leaks, especially from OAGs
with nested subnodes. This will need to be cleaned up in a future iteration
of the code, possibly using the weakref module.