Refilter data upon database update
Currently, _rawdata_windows (rdws) are created via two methods:

1. Slicing/indexing: reducing _rawdata to an rdw via [] notation
2. Filtering: applying a labmda to reduce _rawdata to an rdw

Until now, calls to update() re-indexed the update-refreshed rawdata,
but any previously applied filters were not applied. This could
lead to situations where calls to update() resulted in the OAG
rdw being reset.

This changeset ensures that both indexing and filtering are carried out
when update() is called.

Please note that there is NO guarantee that the rdw will be the same
before and after the update() call as the cached filters could
interact differently with the new _rawdata.

If this behavior is not desirable, refresh from the database (and
associated re-indexing and refiltering) can be turned off by passing
the new norefresh=True flag to update()