RPC discovery needs to be timeout aware as well
This resolves an issue where a dead process leaves a stale entry
in the rpc_discovery table, resulting in rpc.discover() returning true
even though the process hosting the original OAG is long gone.

Since rpc.discover() is supposed to be a read-only function, we are
still relying on the cleanup code in rpc.discoverable to remove the
stale entries.