Overhaul logging
All objects call into the same logger, which is stored in the
gctx(). This changeset effectively makes logging policy consistent
on a per-process basis, and should ameliorate some of the
logger=self.logger soup that seems to pop in the code every now
and then.

Separate OALogs can still be passed into OAGs if logging
different from what is set for the current process is needed.