Rationalize PropProxy.add()
Preserving same logic as before, make PropProxy.add() a much
richer function by making explicit (logically and via documentation)
why functionality is structured the way it is.

Specifically, we make explicit the fact that properties can
be added to an OAG via three paths:

1) Directly by the user on an OAG
2) From a cframe which has been refreshed from a database
3) As a foreign key relationship

Responsibility for cache coherence is also moved directly to the
PropProxy.add() function, as is inter-OAG signalling.

Assorted logging fixes and dead code removal was done as part
of this changeset.