[#42] Add Logging infrastructure
Export a logging-ish device via symbol oalog allowing
for debug, info, warning and error output.

oalog is an alias to OALogManager, which can be called
within a context manager to set an external correlation id.
If set, all log messages will be output with the external
correlation id, allowing for logging coordination with other
services.

Additionally, make some changes to the execution env:

* Make oaenv callable, allowing the elimination of getenv call.

* Make environment fully dot-accessible via pervasive use of
attrdict.

* Clean up logging initialization so that app can override name of
logger when config is merged in, but still log to openarc.core
when used as a standalone library.