[#51] Provide option to query database without throwing
Set throw_on_empty=False for OAG during initialization to ensure
that failed queries to the data layer don't result in exceptions.

The resulting OAG is still iterable (i.e. it has an empty rdf), so
is quite nice from an ergonomics standpoint. So much so that I'm
debating making throw_on_empty True by default...but that's
for another ticket.